Move libraries to backend/lib
