+++ src/phantasm-renderer/backend/lib/stb_image.cc
