+++ src/phantasm-renderer/backend/lib/tiny_obj_loader.cc
